backend/backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
idna==3.10
psycopg2==2.9.9
pyasn1==0.6.1
PyJWT==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
